+++ back/src/post/post.module.ts
@@ -6,11 +6,10 @@ import { Post } from './entities/post.entity';
 import { PostContent } from 'src/post-contents/entities/post-content.entity';
 import { Content } from 'src/contents/entities/content.entity';
 import { Heart } from 'src/hearts/entities/heart.entity';
+import { PostRepository } from './post.repository';
 
 @Module({
-  imports: [
-    TypeOrmModule.forFeature([Post, PostContent, Content, Heart]),
-  ],
+  imports: [TypeOrmModule.forFeature([PostRepository])],
   controllers: [PostController],
   providers: [PostService],
 })
